requirements.txt
requests>=2.31.0
faiss-cpu>=1.7.4
numpy>=1.24.0
urllib3>=2.0.0
tiktoken>=0.5.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
apscheduler>=3.10.4